Track lowest stack remaining seen in vdev_disk as kstat
It may be useful to know if we are being handed an
especially deep stack in vdev_disk_io_start(), rather
than just that we have been called with less than
the threshold remaining.

Additionally update variable names for clarity, notably
reflecting that spl_stack_split_below is not just for vmem
any more.