avoid getting stuck in xnu_alloc_throttled
Previously seen: spl_xat_late_deny increasing, system stalled

Two factors:

Firstly, "usable" memory can go to zero if the kernel
is "only" reclaiming pages when we look (this is not
always a sign of pressure).   We should
always make at least a single PAGE_SIZE allocation
possible, even if the kernel is freeing up memory.

Secondly, we looked for more usable memory than we needed
to meet the allocation once it had reached xat.  Also we
do need ARC to shrink if we end up in xat_late_deny, so
send the signal and let other threads react.