since partial success with delay, loop around delay
The 100ms IOSleep() (which yields the thread, too) had
good success but still was overwhelmed.

Now loop and every 100ms try again, iff the amount
obtained from the system is lower than when we got our
NULL.

We send pressure events every loop so it should not get
stuck infinitey, and our caller has explicitly allowed a
wait on memory.