Handle IOKit device PreferredBlockSize
Get the Preferred Block Size, which may be smaller
than the Physical Block Size, and which appears to
be what is bubbled up to diskutil info -plist
<key>DeviceBlockSize</key>.

In theory, this should only limit the ashift when adding a
vdev or in the readonly ashift vdevprop e.g. zpool get
ashift pool guid), however in practice zpool status -vx
can report mismatches if a smaller-than-physical ashift
was used when adding the vdev intially, and in this case
there can also be vdev.bad_ashift errors that eject disks early
in a scrub (which should not happen).

Therefore return the smaller of the
kIOMediaPreferredBlockSizeKey and kIOPropertyPhysicalBlockSizeKey in
dkimext->dki_pbsize.