Improve cap enforcement in spl_free_wrapper()
spl_free_wrapper() is how ARC knows whether to grow or shrink.

Be a bit more paranoid about many-core allocation bursts
changing values underneat us, in particular not encouraging
ARC to grow immediately after an enforcement on one core.
This should reduce overall enforcements and reduce the
risk of transient crossing of the cap threshold.

If when doing an spl_free_wrapper() we discover that
osif_malloc has grabbed more than the dynamic cap, reduce
the dynamic cap and tell ARC we have a memory shortage.