Handle module_param_call() as well
Linux also uses module_param_call() in places where
ZFS_MODULE_PARAM has yet been changed. This include the
settable aes_impl as similar tunables.

There is also one "module_named_param()" but it is an
old-name alias, so we can ignore it.

Sadly, the Linux _set functions uses sprintf() and not
snprintf(), so we will have to live with warnings until
that can be fixed.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>