taskq_cancel_id() reversed return code
taskq_cancel_id() returns "0" if it definitely managed to
prevent a (delayed) taskq from running (so caller can then
manually clean up), and EBUSY etc if it had to wait for it
to finish, or it ran.

This is used in dmu_objset.c, if it successfully cancelled
a taskq from running ("== 0") it could call rele directly,
causing us to release it twice when we returned the incorrect
return code.

Was triggered by:
tests/functional/upgrade
tests/functional/userquota

Signed-off-by: Jorgen Lundman <lundman@lundman.net>