handle_release() calls list_remove()
handle_alloc_iokit() is called and it needs to bail out at:

    printf("%s found handle after alloc\n", __func__);

This means the call to handle_add() did not add it to the list,
as another node is already there. It then calls
handle_release(lhp); to clean up.

This will eventually end up calling:
       list_remove(&ldi_handle_hash_list[index], lhp);
and panic, as the node isn't on the list.

We cheat a bit by calling list_link_active() to check if
node is "on the list".

Triggered by:

tests/functional/cli_root/zpool_add (test 008 or 009)