nerfing and clarity around kmem_cache_reap_now()
Fix kmem_cache_reap_active() for macOS.  The previous
return of B_FALSE just followed the FreeBSD practice for
their kmem-related SPL code (which is very different from
ours), but macOS's reaping of empty is gated by the
kmem_reaping flag.

Linux and FreeBSD have both softened kmem_cache_reap_now()
because it can be called a lot and highly concurrently
from common code in arc.c.  We do the same here, just
calling kmem_reap() instead of reaping a particular
cache *RIGHT NOW*.  The open question is whether omitting
the zeroing of the working set for the cache in question
is the right thing to do.

An alternative approach if this one isn't effective could
involve bailing out for a particular cache lower down in
the stack, e.g. if cache_reap_lock is held.