deal with a possible data race in vmem_update() thread (#13)
* The new hash table's size is calculated, and an allocation
based on that size is made outside the arena lock
immediately after a cv_broadcast() is called.  The
allocation cannot be made under a lock.

Previously there was some recalculation once the lock had
been taken; but the allocation size calculated once the
lock is obtained could differ from what was calculated
unlocked.  Free the allocated memory and return without
rescaling the hash table if that's the case.

* small clarification of vmem_bucket_wake_all_waiters()

* small cleanup in spl_free_set_pressure()

* broadcast-> lock; signal; unlock

There's only one possible waiter, and we should wait until
it's actually waiting before signalling it.

* reduce kmem_reap_preempt_bytes to what it was in the old times
Signed-off-by: Jorgen Lundman <lundman@lundman.net>