Apple Silicon sequential consistency and vmem thread-call
* Take care in not letting another thread stomp the
returned memory address obtained from the worker thread in
vmem_alloc_in_worker_thread()

* Fixed an off-by-a-factor-of-1024 when creating the
abd kmem caches.   Apple Silicon machines with 8GB+ RAM
will now use larger slab sizes in the kmem caches, for the
performance gain that was nerfed by a 4096 rather than 4.

* Drop VM_FIRSTFIT in the creation of the abd vmem caches.
The default ("instant fit") has acceptable properties even
with the larger quanta it obtains from the bucket heap, and
allocates in constant time (and does less work).

MORE:

Chasing down an apparent early return of mwait() affecting
the cv_timedwait() in vmem_alloc_in_worker_thread(), I
made some improvements for Apple Silicon:

* Add some barriers to enforce total order / sequential
consistency on Apple Silicon in mutex and condvar
code (the compiler will add some "dmb ish", which is a
data memory barrier for inner shareablility domain,
which is [a] default that the Xcode compiler (and
clang-llvm) uses, and is the assembly used in Apple code
except for updating the global clock and [b] probably
right, assuming that the Ultra chips are both in the
"innner", or that the Ultra does the right thing with "dmb
ish".

See:
https://community.arm.com/arm-community-blogs/b/architectures-and-processors-blog/posts/memory-access-ordering-part-3---memory-access-ordering-in-the-arm-architecture

* Reworked the thread_call-related functions in spl-vmem.c

Rather than wrapping macros around non-atomic parts of a
structure, declared them as _Atomic, and used the plain
syntax.

Added some barriers by hand.

The above has negligible performance impact on Apple
Silicon, and does nothing on x86_64.

kpreempt() to perhaps avoid the cv_timedwait() entirely.

* cv_timedwait* debugging: assert msleep_return is 0

Returning 1 means that we have been sent a cv_signal()
successfully, rather than a condition like a
timeout (we'd return -1) or an interrupting signal (we'd
return 0).   But this implies that msleep_result is 0,
which perhaps it isn't.