TIMESHARE threads by default, but allow disabling
Don't TIMESHARE for very low priority threads, because we
don't want their priorities to rise (they can get very
high on an unbusy system and with certain "good-behaviour"
patterns), or for strict-FIFO taskqs (since the dynamic
behaviour of the immediately previous task might leaad to
a poor prediction of the current task's needs).

Mutex debugging thread should not timeshare.