Rework kernel thread startup
Previously we had the kernel start a thread which would
race ahead during which the parent would set attributes
such as thread importance (offset from xnu KERN_BASEPRI
priority 81), throughput and latency QOS, and the
TIMESHARE attribute.  Doing this within the parent was
unreliable thanks to a race, and adjusting within the
child thread after certain events (such as any IOKit work)
also didn't always work.  This was most obvious when
threads had timeshare turned off (so that they would not
be promoted to high kernel priorities by xnu during what
it considers to be good behaviour).

Thread creation now uses a protocol wherein the parent
fills in and passes a pointer to a struct containing
information it needs to set the thread up and call the
worker function.  The child will wait against a xnu
mutex (we cannot use spl-mutex.c mutexes early on) until
the parent is ready, then the parent will loop doing xnu
msleep() until it sees that the child has set the done
flag.  The parent's msleep lets the child acquire the
mutex and set up various thread attributes.  It will then
set a flag and lck->wakeup_one->unlock then jump to the
worker function.  The wakeup and unlock lets the parent
see the child's done flag, and the parent will return.

Making the parent wait for the child allows us to use auto
variables (or procede to a free()) to set up various
structures which are passed to the new thread creation
functions as pointers.  The parent will not return to its
calling function before the child no longer can use those
structures (as the pointers are not passed to the worker
function).

When it becomes necessary it will be straightforward to
pass more information from the parent to the child, and
the child could pass more information to the parent.

The primary function
spl_thread_create_named_with_extpol_and_qos() accepts
timeshare and qos policy arguments, which may be NULL.  A
wrapper function passes in default values (latency and
througput QOS being NULLs), and that in turn is wrapped by
the macros in thread.h for the benefit of openzfs common
source.

Mutex debugging thread should not timeshare (we don't want
it demoted from high priority during a period where it is
busy dealing with the list), so it uses the new unwrapped
API, spl_thread_create_named_with_extpol_and_qos().

spl-taskq.c has been reworked to use
spl_thread_create_named_with_extpol_and_qos() too.

Throughput and latency QOS is now set more conservatively,
which makes analysing per-thread priority ranges in
spindump easier.  Threads with lower than minclsyspri have
no QOS policy set, and timeshare is disabled; this should
stop low priority scrub threads from being promoted to
very high priority.