zfs-tests: support newer macOS readOnly root
Since BigSur days, "/" on macOS is a readOnly, sealed
archive which means zfs-tests can not mount the expected
/testpool (which all scripts assume are at root). We can
work around this by leveraging /etc/synthetic.conf to
create "virtual directories". They can be mounted on, but
not modified. This way *most* pool names are accounted
for in zfs-tests (more will come).

We also have to make log_must(rm /TESTPOOL) be optional.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>