Export before removing file
Due to a macOS quirk, we have to close the files
in a filebased pool during some operations, which means
the export -f task will fail to open the vdevs again, the
pool will be suspended and zpool_export_004_pos will
hang forever.

By deleting the file after export, everything works as
expected.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>