zpool freeze can hang in spa_evicting_os_wait
It appears upstream can still unmount, and export
the pool - to later re-import to unfreeze. However
on macOS, export will hang in spa_evicting_os_wait().

It is unclear why we hang, but it could be we go through
more syncs during unmount, and that spa_freeze_txg is set to
txg + TXG_SIZE

Now we clear (restoring UINT64_MAX) spa_freeze_txg in the
unmount() call, ensuring unmount and export works.

We might not have identical "zpool freeze" as upstream, but
as the zpool sources say:

 * 'freeze' is a vile debugging abomination, so we treat
 * it as such.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>